Server without friends

+++ controllers/pictures.js
@@ -16,10 +16,13 @@ module.exports.create = async function(req, res) {
           text: req.body.text,
           parent: req.params.parentID,
           p_sort: maxSort + 1,
-          boysGreyPicture: req.files['boysGreyPicture'] ? req.files['boysGreyPicture'][0].path : '',
-          girlsGreyPicture: req.files['girlsGreyPicture'] ? req.files['girlsGreyPicture'][0].path : '',
-          boysColorPicture: req.files['boysColorPicture'] ? req.files['boysColorPicture'][0].path : '',
-          girlsColorPicture: req.files['girlsColorPicture'] ? req.files['girlsColorPicture'][0].path : ''
+          boysGreyPicture: req.files ? (req.files['boysGreyPicture'] ? req.files['boysGreyPicture'][0].path : '') : '',
+          girlsGreyPicture: req.files ? (req.files['girlsGreyPicture'] ? req.files['girlsGreyPicture'][0].path : '') : '',
+          boysColorPicture: req.files ? (req.files['boysColorPicture'] ? req.files['boysColorPicture'][0].path : '') : '',
+          girlsColorPicture: req.files ? (req.files['girlsColorPicture'] ? req.files['girlsColorPicture'][0].path : '') : '',
+          invisible: req.body.invisible,
+          system: req.body.system,
+          user: req.user.id
         }).save()
         res.status(201).json(picture)
       } catch (e) {
@@ -89,16 +92,23 @@ module.exports.remove = async function(req, res) {
 
     await Picture.updateMany(pictures, {$set: {p_sort: p_sort - 1}})
     */
-    
-    if (deletePicture.folder === true) {
-      message = 'Папка удалена.'
-    } else {
-      message = 'Картинка удалена.'
+    if (deletePicture.system === true) {
+      res.status(200).json({
+        message: 'Этот объект нельзя удалить.'
+      })  
     }
-    await Picture.deleteOne(deletePicture)
-    res.status(200).json({
-      message: message
+    else {
+      if (deletePicture.folder === true) {
+        message = 'Папка удалена.'
+      } 
+      else {
+        message = 'Картинка удалена.'
+      }
+        await Picture.deleteOne(deletePicture)
+        res.status(200).json({
+          message: message
     })
+  }
   } catch (e) {
     errorHandler(res, e)
   }
